pyproject.toml
[tool.poetry]
name = "rtk-it"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.44"
pydantic-settings = "^2.11.0"
fastapi = "^0.119.1"
uvicorn = "^0.38.0"
requests = "^2.32.5"
bcrypt = "^5.0.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
pydantic = {extras = ["email"], version = "^2.12.3"}
python-multipart = "^0.0.20"
httpx = "^0.28.1"
pandas = "^2.3.3"
websockets = "^15.0.1"
redis = "^7.0.1"
fastapi-cache2 = "^0.2.2"
tiktoken = "^0.12.0"
asyncpg = "^0.30.0"
ruff = "^0.14.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = false